test(scope): cast the enum ordinal to satisfy MISRA C:2012 Rule 10.3
CI's MISRA leg caught what the pre-push gate does not run:

  issue-1117-depth-ambiguity.test.c:56:27: misra violation [misra-c2012-10.3]

`u8 secondOrdinal <- second;` assigns an enum straight to u8, crossing
essential type categories. Rule 10.3 forbids that. Introduced by the ordinal
assertions added in the previous commit — the review suggestion was verified to
transpile and execute, but not against MISRA.

Only the `main`-scope assignment was flagged; the same conversion inside
`A.firstOrdinal()` (`return current;` with a u8 return type) is not, so only this
one site needed the cast.

Verified: `npm run validate:c` now passes all five tools (cppcheck C, cppcheck
C++, clang-tidy, MISRA, flawfinder).

Refs #1117

+++ tests/scope/issue-1117-depth-ambiguity.expected.cpp
@@ -53,7 +53,7 @@ int main(void) {
     uint8_t inScopeOrdinal = A__firstOrdinal();
     if (inScopeOrdinal != 0) return 3;
     A__B second = A__B__d;
-    uint8_t secondOrdinal = second;
+    uint8_t secondOrdinal = static_cast<uint8_t>(second);
     if (secondOrdinal != 1) return 4;
     if (A__B_c != 7) return 5;
     return 0;
